# backend/src/main/resources/swagger.yaml
openapi: 3.0.4
info:
  title: Session Auth API - Document Management
  description: API for managing documents in the Session Authentication system
  version: 1.0.0
  contact:
    email: support@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.example.com/api/v1
    description: Production server
...
paths:
...
  /scholarships/create:
    post:
      tags:
        - Scholarship
      summary: Create a new scholarship
      description: Creates a new scholarship with the provided details. Requires EMPLOYEE authority.
      operationId: createScholarship
      security:
        - bearerAuth: [EMPLOYEE]
      requestBody:
        description: Scholarship to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScholarshipDTO'
      responses:
        '201':
          description: Scholarship created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScholarshipDTO'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /scholarships/{id}:
    get:
      tags:
        - Scholarship
      summary: Get scholarship by ID
      description: Returns a single scholarship by its ID. Requires EMPLOYEE authority.
      operationId: getScholarshipById
      security:
        - bearerAuth: [EMPLOYEE]
      parameters:
        - name: id
          in: path
          description: ID of scholarship to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScholarshipDTO'
        '400':
          description: Invalid ID supplied
        '404':
          description: Scholarship not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Scholarship
      summary: Update an existing scholarship
      description: Updates an existing scholarship. Requires EMPLOYEE authority.
      operationId: updateScholarship
      security:
        - bearerAuth: [EMPLOYEE]
      parameters:
        - name: id
          in: path
          description: ID of scholarship to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Scholarship data to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScholarshipDTO'
      responses:
        '200':
          description: Scholarship updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScholarshipDTO'
        '400':
          description: Invalid ID or input
        '404':
          description: Scholarship not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Scholarship
      summary: Delete a scholarship
      description: Deletes a scholarship by ID. Requires EMPLOYEE authority.
      operationId: deleteScholarship
      security:
        - bearerAuth: [EMPLOYEE]
      parameters:
        - name: id
          in: path
          description: ID of scholarship to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Scholarship deleted successfully
        '400':
          description: Invalid ID supplied
        '404':
          description: Scholarship not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /scholarships/getAll:
    get:
      tags:
        - Scholarship
      summary: Get all scholarships
      description: Returns a list of all scholarships. Requires EMPLOYEE authority.
      operationId: getAllScholarships
      security:
        - bearerAuth: [EMPLOYEE]
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScholarshipDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /job-titles/create:
    post:
      tags:
        - JobTitle
      summary: Create a new job title
      description: Creates a new job title with the provided details. Requires ADMIN authority.
      operationId: createJobTitle
      security:
        - bearerAuth: [ADMIN]
      requestBody:
        description: Job title to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobTitleDTO'
      responses:
        '201':
          description: Job title created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobTitleDTO'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /job-titles/{id}:
    get:
      tags:
        - JobTitle
      summary: Get job title by ID
      description: Returns a single job title by its ID. Requires ADMIN authority.
      operationId: getJobTitleById
      security:
        - bearerAuth: [ADMIN]
      parameters:
        - name: id
          in: path
          description: ID of job title to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobTitleDTO'
        '400':
          description: Invalid ID supplied
        '404':
          description: Job title not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - JobTitle
      summary: Update an existing job title
      description: Updates an existing job title. Requires ADMIN authority.
      operationId: updateJobTitle
      security:
        - bearerAuth: [ADMIN]
      parameters:
        - name: id
          in: path
          description: ID of job title to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Job title data to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobTitleDTO'
      responses:
        '200':
          description: Job title updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobTitleDTO'
        '400':
          description: Invalid ID or input
        '404':
          description: Job title not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - JobTitle
      summary: Delete a job title
      description: Deletes a job title by ID. Requires ADMIN authority.
      operationId: deleteJobTitle
      security:
        - bearerAuth: [ADMIN]
      parameters:
        - name: id
          in: path
          description: ID of job title to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Job title deleted successfully
        '400':
          description: Invalid ID supplied
        '404':
          description: Job title not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /job-titles/getAll:
    get:
      tags:
        - JobTitle
      summary: Get all job titles
      description: Returns a list of all job titles. Requires ADMIN authority.
      operationId: getAllJobTitles
      security:
        - bearerAuth: [ADMIN]
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobTitleDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /kafedras/create:
    post:
      tags:
        - Kafedra
      summary: Create a new kafedra
      description: Creates a new kafedra (department) with the provided details. Requires EMPLOYEE authority.
      operationId: createKafedra
      security:
        - bearerAuth: [EMPLOYEE]
      requestBody:
        description: Kafedra to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KafedraDTO'
      responses:
        '201':
          description: Kafedra created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KafedraDTO'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /kafedras/{id}:
    get:
      tags:
        - Kafedra
      summary: Get kafedra by ID
      description: Returns a single kafedra by its ID. Requires EMPLOYEE authority.
      operationId: getKafedraById
      security:
        - bearerAuth: [EMPLOYEE]
      parameters:
        - name: id
          in: path
          description: ID of kafedra to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KafedraDTO'
        '400':
          description: Invalid ID supplied
        '404':
          description: Kafedra not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Kafedra
      summary: Update an existing kafedra
      description: Updates an existing kafedra. Requires EMPLOYEE authority.
      operationId: updateKafedra
      security:
        - bearerAuth: [EMPLOYEE]
      parameters:
        - name: id
          in: path
          description: ID of kafedra to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Kafedra data to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KafedraDTO'
      responses:
        '200':
          description: Kafedra updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KafedraDTO'
        '400':
          description: Invalid ID or input
        '404':
          description: Kafedra not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Kafedra
      summary: Delete a kafedra
      description: Deletes a kafedra by ID. Requires EMPLOYEE authority.
      operationId: deleteKafedra
      security:
        - bearerAuth: [EMPLOYEE]
      parameters:
        - name: id
          in: path
          description: ID of kafedra to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Kafedra deleted successfully
        '400':
          description: Invalid ID supplied
        '404':
          description: Kafedra not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /kafedras/getAll:
    get:
      tags:
        - Kafedra
      summary: Get all kafedras
      description: Returns a list of all kafedras. Requires EMPLOYEE authority.
      operationId: getAllKafedras
      security:
        - bearerAuth: [EMPLOYEE]
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KafedraDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
...
    ScholarshipDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
          description: Scholarship ID
        name:
          type: string
          example: "Academic Excellence Scholarship"
          description: Name of the scholarship
        amount:
          type: string
          example: "$5000"
          description: Amount awarded by the scholarship
        scholarshipType:
          type: string
          enum:
            - ACADEMIC
            - SOCIAL
            - PRESIDENT
          example: "ACADEMIC"
          description: Type of the scholarship (ACADEMIC, SOCIAL, or PRESIDENT)
      required:
        - name
        - amount
        - scholarshipType
...
    JobTitleDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
          readOnly: true
          description: Job title ID
        name:
          type: string
          example: "Software Engineer"
          description: Name of the job title
        description:
          type: string
          example: "Develops software applications"
          description: Description of the job title
          nullable: true
        subdivisionId:
          type: integer
          format: int64
          example: 5
          description: ID of the associated subdivision
      required:
        - name
        - subdivisionId
    KafedraDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
          readOnly: true
          description: Kafedra ID
        name:
          type: string
          example: "Computer Science Department"
          description: Name of the kafedra
        email:
          type: string
          format: email
          example: "cs@university.edu"
          description: Contact email
        room:
          type: string
          example: "Room 101"
          description: Room number/location
        phone:
          type: string
          example: "+1234567890"
          description: Contact phone number
        credentialsNonExpired:
          type: boolean
          example: true
          description: Indicates if credentials are non-expired
        instituteId:
          type: integer
          format: int64
          example: 1
          description: ID of the associated institute
      required:
        - name
        - email
        - phone
        - instituteId
    Error:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2023-10-01T12:00:00Z"
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Invalid input data"
        path:
          type: string
          example: "/api/v1/documents"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication